rename utils script under test to avoid ambiguity
