+++ tests/tools/test_dist_part.py
@@ -25,7 +25,7 @@ from tools.verification_utils import (
     verify_partition_data_types,
     verify_partition_formats,
 )
-from utils import create_chunked_dataset
+from pytest_utils import create_chunked_dataset
 
 
 def _test_chunk_graph(
